remove import of Base class for test_timedelta.py
