+++ pandas/tests/resample/test_base.py
@@ -12,7 +12,7 @@ import pandas as pd
 from pandas import DataFrame, Series
 from pandas.core.groupby.groupby import DataError
 from pandas.core.indexes.period import PeriodIndex
-from pandas.core.indexes.timedeltas import TimedeltaIndex
+from pandas.core.indexes.timedeltas import TimedeltaIndex, timedelta_range
 from pandas.core.resample import TimeGrouper
 import pandas.util.testing as tm
 from pandas.util.testing import (
@@ -229,3 +229,25 @@ class Base(object):
         result = s.resample(freq).quantile(q)
         expected = s.resample(freq).agg(lambda x: x.quantile(q))
         tm.assert_series_equal(result, expected)
+
+
+class TestTimedeltaIndex(Base):
+    _index_factory = lambda x: timedelta_range
+
+    @pytest.fixture
+    def _index_start(self):
+        return '1 day'
+
+    @pytest.fixture
+    def _index_end(self):
+        return '10 day'
+
+    @pytest.fixture
+    def _series_name(self):
+        return 'tdi'
+
+    def create_series(self):
+        i = timedelta_range('1 day',
+                            '10 day', freq='D')
+
+        return Series(np.arange(len(i)), index=i, name='tdi')
